feat: Enhance analysis options and update dependencies
- Updated `analysis_options.yaml` to exclude generated files from analysis.
- Updated `pubspec.yaml` and `pubspec.lock` to include the latest dependencies.
- Refactored various components and models for improved functionality and consistency.
- Introduced new BLoC files for connectivity and notification management.
- Cleaned up unused imports and removed deprecated files to streamline the codebase.

+++ lib/features/read_quran/presentation/view/widgets/top_title_widget.dart
@@ -33,7 +33,7 @@ class TopTitleWidget extends StatelessWidget {
                   },
                   child: bookmarkIcon(
                     context,
-                    height: context.customOrientation(35.h, 55.h),
+                    height: context.customOrientation(35.h, 55.h) as double,
                     pageNum: lastPageRead,
                   ),
                 );
@@ -43,7 +43,7 @@ class TopTitleWidget extends StatelessWidget {
             Text(
               '${'juz'}: ${convertNumbers(quranCtrl.getJuzByPage(index).juz.toString())}',
               style: TextStyle(
-                fontSize: context.customOrientation(18.0, 22.0),
+                fontSize: context.customOrientation(18.0, 22.0) as double,
                 fontFamily: 'naskh',
                 color: const Color(0xff77554B),
               ),
@@ -52,7 +52,7 @@ class TopTitleWidget extends StatelessWidget {
             Text(
               quranCtrl.getSurahNameFromPage(index),
               style: TextStyle(
-                fontSize: context.customOrientation(18.0, 22.0),
+                fontSize: context.customOrientation(18.0, 22.0) as double,
                 // fontWeight: FontWeight.bold,
                 fontFamily: 'naskh',
                 color: const Color(0xff77554B),
